rename hooks and move hooks into hooks folder
